root初期化関数追加
